requirements.txt
pysrt
python-dotenv
python-telegram-bot
jinja2
